# Optimize parsing option args in load