@Watson1978 2023-05-06
This is similar way with https://github.com/ohler55/oj/pull/683